Replace non free Ubuntu font
This may come as a surprise to some but after careful review the Ubuntu font
has been deemed non-free by both Debian and Fedora, see:
https://bugs.launchpad.net/ubuntu-font-licence/+bug/769874
https://bugs.launchpad.net/ubuntu-font-family/+bug/1211154

Unfortunately replacing it takes 2 fonts. The free comfortaa font is a
decent replacement style wise, but has non full cyrillic coverage. Possibly
its cyrillic coverage is good enough, but it is better to be safe.

So this commits replaces the Ubuntu font with Comfortaa for Latin character
using languages, and adds good old DejaVu for use with cyrillic languages.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>